Put claim cleaners in their own namespace
All the cleaners are now based on `Claim::Cleaner`, which is currently empty
but can be used for common code.
