+++ app/models/claim/concerns/cleaner.rb
@@ -1,0 +1,4 @@
+module Claim
+  class Cleaner
+  end
+end
